requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
gitpython
docker
